+++ scripts/ci/pylint_todo.txt
@@ -61,7 +61,7 @@
 ./airflow/providers/amazon/aws/operators/sagemaker_model.py
 ./airflow/providers/amazon/aws/operators/sagemaker_training.py
 ./airflow/providers/amazon/aws/operators/sagemaker_transform.py
-./airflow/contrib/operators/sagemaker_tuning_operator.py
+./airflow/providers/amazon/aws/operators/sagemaker_tuning.py
 ./airflow/contrib/operators/segment_track_event_operator.py
 ./airflow/contrib/operators/sftp_to_s3_operator.py
 ./airflow/contrib/operators/slack_webhook_operator.py
